[feature/weather_list] - Weather List ADD: * Constants.kt
UPT:
* Fetch groups integration on WeatherGroupFragment.kt
* Set WeatherGroupView.kt bg base on data temperature
* Show SkeletonLayout while data is loading

+++ app/src/main/java/com/example/hcassessment/core/data/pojo/group/WeatherGroup.kt
@@ -47,6 +47,6 @@ data class Weather(
 )
 
 data class Wind(
-    val deg: Int,
-    val speed: Int
+    val deg: Double,
+    val speed: Double
 )
